mcp23x17: convert mcp23017 to a generic driver over I2C and SPI version
